+++ src/clj/withings_client/users.clj
@@ -28,6 +28,10 @@
   (log/info "users-list")
   (db/get-users))
 
+(defn valid-users
+  []
+  (db/valid-users))
+
 ;; ;; no. update-user-by-name ではないと使い道がない。
 ;; (defn update-cid!
 ;;   [params]
@@ -38,15 +42,21 @@
   (db/user-by-name {:name name}))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; no use?
+
 (defn update-tokens-by-name!
+  "use from tokens/fetch-and-store!
+   first insert user name, cid, secret,
+   then updte the record with access, refresh token using key `name`"
   [params]
   (db/update-tokens-by-name! params))
 
-(defn update-tokens-by-userid!
+(defn update-tokens!
+  "Update id's access-token and refresh-token.
+   Sometimes, refresh-token is not updated. So,
+   return value is 1 or 2 when update successed."
   [params]
-  (log/info "update-tokens-by-userid!" params)
-  (db/update-tokens-by-userid! params))
+  (log/info "update-tokens!" params)
+  (db/update-tokens! params))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 (defn toggle-valid!